# 01-basic-auth/go/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "01 - Basic Auth (Go)",
    "version": "1.0.0",
    "description": "Аутентификация по email + паролю. Все операции принимают email и пароль в теле запроса (JSON)."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" }
                  },
                  "required": ["status"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "summary": "Регистрация нового пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь создан",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "description": "Bad Request — пустые поля" },
          "409": { "description": "Пользователь с таким email уже существует" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Вход по email и паролю",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Неверный email или пароль" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/v1/auth/delete": {
      "delete": {
        "summary": "Удаление аккаунта",
        "description": "Принимает email и пароль в теле запроса — проверяет личность и удаляет аккаунт.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь удалён",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Пользователь успешно удалён" }
                  },
                  "required": ["message"]
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Неверный email или пароль" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "ivanov@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "example": "1234"
          }
        },
        "required": ["email", "password"],
        "example": {
          "email": "ivanov@example.com",
          "password": "1234"
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890" },
          "email": { "type": "string", "format": "email", "example": "ivanov@example.com" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-03-10T12:00:00Z" }
        },
        "required": ["id", "email", "created_at"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Пользователь успешно зарегистрирован" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["message", "user"]
      }
    }
  }
}